+++ relm4/src/lib.rs
@@ -1,10 +1,10 @@
 //! An idiomatic GUI library inspired by Elm and based on gtk4-rs.
 
 #![doc(
-    html_logo_url = "https://raw.githubusercontent.com/AaronErhardt/relm4/main/assets/Relm_logo.svg"
+    html_logo_url = "https://raw.githubusercontent.com/Relm4/Relm4/main/assets/Relm_logo.svg"
 )]
 #![doc(
-    html_favicon_url = "https://raw.githubusercontent.com/AaronErhardt/relm4/main/assets/Relm_logo.svg"
+    html_favicon_url = "https://raw.githubusercontent.com/Relm4/Relm4/main/assets/Relm_logo.svg"
 )]
 #![warn(
     missing_debug_implementations,